Create apod-html
This creates a simplified version of the apod site that you can embed in another html file on your website. It will display only the picture itself and the credit. I do not run it as a cgi script on the website itself as it's unnecessary for it to execute more than once a day (and I have, like, so little RAM available, it's pathetic). I run it via cronjob at what I estimate is about midnight where the apod website is hosted.
* 6 * * * bash /path/to/script > /dev/null 2>&1