Create apod-html.sh
This creates a tiny version of the apod, including only the image itself and the owner. It can be embedded on another page as a little card. Though I suppose you could refunction it as a cgi script, it isn't necessary for it to run more than once a day, which is why I've got it running once a day via cronjob during what I imagine must be about midnight where the apod server is located.
* 6 * * * bash /pathe/to/script > /dev/null 2>&1